backend_server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
drf-spectacular==0.26.5
psycopg2-binary==2.9.9
Pillow==10.1.0
python-dotenv==1.0.0
